created header & scrapped Nav
